Merging David's fixes into the latest master.
Fix for a problem where the SimpleMainView gets added with nothing when a tabbed view is what the developers want to use as mainview, which causes an ugly white flash between splash screen and app display. Added a timer-based splash removal. Not as nice as an onFinish, but I was able to get it to work. Left the other code commented out. Maybe onFinish can call the new WaitThenDo task class?

Conflicts:

	platform/android/Rhodes/src/com/rhomobile/rhodes/RhodesActivity.java
	platform/android/Rhodes/src/com/rhomobile/rhodes/SplashScreen.java